Refactor chat UI functions to separate module
